Stereo context switching fixes
